requirements.txt
flask
requests
python-dotenv
gunicorn
python-dateutil
openai
schedule
flagsmith